feat: Implement tests for decrement click button
In this commit, the decrement button's onClick event is checked to
see if it decrement's the value of count accordingly i.e by 1.
when count is 0 and the decrement button is clicked, we check to
see if there's a -1 value in the document.
The test fails if there is 0 in the document after the decrement button a

has been clicked